+++ tests/sampling/discrete/test_sample_int.py
@@ -22,31 +22,31 @@ def get_probabilities(n: int) -> np.ndarray[float]:
 def test_sample_int_argument_validation(use_numba: bool):
     # --- n < 1 ------------------------------------------
     with pytest.raises(ValueError):
-        sample_int(n=0, accelerated=use_numba)
+        sample_int(n=0, use_numba=use_numba)
 
     with pytest.raises(ValueError):
-        sample_int(n=-10, accelerated=use_numba)
+        sample_int(n=-10, use_numba=use_numba)
 
     # --- k < 1 ------------------------------------------
     with pytest.raises(ValueError):
-        sample_int(n=10, k=0, accelerated=use_numba)
+        sample_int(n=10, k=0, use_numba=use_numba)
 
     with pytest.raises(ValueError):
-        sample_int(n=10, k=-5, accelerated=use_numba)
+        sample_int(n=10, k=-5, use_numba=use_numba)
 
     # --- k > n when replace=False -----------------------
     with pytest.raises(ValueError):
-        sample_int(n=10, k=20, replace=False, accelerated=use_numba)
+        sample_int(n=10, k=20, replace=False, use_numba=use_numba)
 
     with pytest.raises(ValueError):
-        sample_int(n=100, k=101, replace=False, accelerated=use_numba)
+        sample_int(n=100, k=101, replace=False, use_numba=use_numba)
 
     # --- p invalid shape --------------------------------
     with pytest.raises(ValueError):
-        sample_int(n=10, k=5, replace=True, p=np.array([0.1, 0.9]), accelerated=use_numba)  # wrong size
+        sample_int(n=10, k=5, replace=True, p=np.array([0.1, 0.9]), use_numba=use_numba)  # wrong size
 
     with pytest.raises(ValueError):
-        sample_int(n=4, k=2, replace=True, p=np.array([[0.1, 0.4], [0.4, 0.1]]), accelerated=use_numba)  # wrong shape
+        sample_int(n=4, k=2, replace=True, p=np.array([[0.1, 0.4], [0.4, 0.1]]), use_numba=use_numba)  # wrong shape
 
 
 # =================================================================================================
@@ -54,9 +54,10 @@ def test_sample_int_argument_validation(use_numba: bool):
 # =================================================================================================
 @pytest.mark.parametrize("n", [10, 100, 1000, 10000])
 @pytest.mark.parametrize("use_numba", [True, False])
-def test_sample_int_uniform_with_replacement_invariants_single(n: int, use_numba: bool):
+@pytest.mark.parametrize("p", [None, np.zeros(0, np.float64)])
+def test_sample_int_uniform_with_replacement_invariants_single(n: int, use_numba: bool, p: np.ndarray | None):
     # --- act ---------------------------------------------
-    sample = sample_int(n, replace=True, accelerated=use_numba)
+    sample = sample_int(n, replace=True, p=p, use_numba=use_numba)
 
     # --- assert ------------------------------------------
     assert isinstance(sample, numbers.Integral)
@@ -77,9 +78,10 @@ def test_sample_int_uniform_with_replacement_invariants_single(n: int, use_numba
     ],
 )
 @pytest.mark.parametrize("use_numba", [True, False])
-def test_sample_int_uniform_with_replacement_invariants_multiple(n: int, k: int, use_numba: bool):
+@pytest.mark.parametrize("p", [None, np.zeros(0, np.float64)])
+def test_sample_int_uniform_with_replacement_invariants_multiple(n: int, k: int, use_numba: bool, p: np.ndarray | None):
     # --- act ---------------------------------------------
-    samples = sample_int(n, k, replace=True, accelerated=use_numba)
+    samples = sample_int(n, k, replace=True, p=p, use_numba=use_numba)
 
     # --- assert ------------------------------------------
     assert isinstance(samples, np.ndarray)
@@ -89,9 +91,10 @@ def test_sample_int_uniform_with_replacement_invariants_multiple(n: int, k: int,
 
 
 @pytest.mark.parametrize("use_numba", [True, False])
-def test_sample_int_uniform_with_replacement_duplicates(use_numba: bool):
+@pytest.mark.parametrize("p", [None, np.zeros(0, np.float64)])
+def test_sample_int_uniform_with_replacement_duplicates(use_numba: bool, p: np.ndarray | None):
     # --- act ---------------------------------------------
-    samples = sample_int(n=1000, k=900, replace=True, accelerated=use_numba)  # very unlikely all are unique
+    samples = sample_int(n=1000, k=900, replace=True, p=p, use_numba=use_numba)  # very unlikely all are unique
 
     # --- assert ------------------------------------------
     assert len(set(samples)) < 900  # there are duplicates, which is expected with replacement
@@ -111,11 +114,13 @@ def test_sample_int_uniform_with_replacement_duplicates(use_numba: bool):
     ],
 )
 @pytest.mark.parametrize("use_numba", [True, False])
-def test_sample_int_uniform_with_replacement_seed(n: int, k: int, use_numba: bool):
+@pytest.mark.parametrize("p", [None, np.zeros(0, np.float64)])
+def test_sample_int_uniform_with_replacement_seed(n: int, k: int, use_numba: bool, p: np.ndarray | None):
     # --- act ---------------------------------------------
-    samples_1 = sample_int(n, k, replace=True, accelerated=use_numba, seed=123)
-    samples_2 = sample_int(n, k, replace=True, accelerated=use_numba, seed=123)
-    samples_3 = sample_int(n, k, replace=True, accelerated=use_numba, seed=None)
+    samples_1 = sample_int(n, k, replace=True, p=p, use_numba=use_numba, seed=123)
+    samples_2 = sample_int(n, k, replace=True, p=p, use_numba=use_numba, seed=123)
+    samples_3 = sample_int(n, k, replace=True, p=p, use_numba=use_numba, seed=None)
+    samples_4 = sample_int(n, k, replace=True, p=p, use_numba=use_numba, seed=0)
 
     # --- assert ------------------------------------------
     np.testing.assert_array_equal(samples_1, samples_2)
@@ -123,6 +128,7 @@ def test_sample_int_uniform_with_replacement_seed(n: int, k: int, use_numba: boo
         # in this case, it's very unlikely that samples_3 matches samples_1 or samples_2
         assert not list(samples_1) == list(samples_3)
         assert not list(samples_2) == list(samples_3)
+        assert not list(samples_3) == list(samples_4)
 
 
 # =================================================================================================
@@ -130,9 +136,10 @@ def test_sample_int_uniform_with_replacement_seed(n: int, k: int, use_numba: boo
 # =================================================================================================
 @pytest.mark.parametrize("n", [10, 100, 1000, 10000])
 @pytest.mark.parametrize("use_numba", [True, False])
-def test_sample_int_uniform_without_replacement_invariants_single(n: int, use_numba: bool):
+@pytest.mark.parametrize("p", [None, np.zeros(0, np.float64)])
+def test_sample_int_uniform_without_replacement_invariants_single(n: int, use_numba: bool, p: np.ndarray | None):
     # --- act ---------------------------------------------
-    sample = sample_int(n, replace=False, accelerated=use_numba)
+    sample = sample_int(n, replace=False, p=p, use_numba=use_numba)
 
     # --- assert ------------------------------------------
     assert isinstance(sample, numbers.Integral)
@@ -152,9 +159,12 @@ def test_sample_int_uniform_without_replacement_invariants_single(n: int, use_nu
     ],
 )
 @pytest.mark.parametrize("use_numba", [True, False])
-def test_sample_int_uniform_without_replacement_invariants_multiple(n: int, k: int, use_numba: bool):
+@pytest.mark.parametrize("p", [None, np.zeros(0, np.float64)])
+def test_sample_int_uniform_without_replacement_invariants_multiple(
+    n: int, k: int, use_numba: bool, p: np.ndarray | None
+):
     # --- act ---------------------------------------------
-    samples = sample_int(n, k, replace=False, accelerated=use_numba)
+    samples = sample_int(n, k, replace=False, p=p, use_numba=use_numba)
 
     # --- assert ------------------------------------------
     assert isinstance(samples, np.ndarray)
@@ -177,11 +187,13 @@ def test_sample_int_uniform_without_replacement_invariants_multiple(n: int, k: i
     ],
 )
 @pytest.mark.parametrize("use_numba", [True, False])
-def test_sample_int_uniform_without_replacement_seed(n: int, k: int, use_numba: bool):
+@pytest.mark.parametrize("p", [None, np.zeros(0, np.float64)])
+def test_sample_int_uniform_without_replacement_seed(n: int, k: int, use_numba: bool, p: np.ndarray | None):
     # --- act ---------------------------------------------
-    samples_1 = sample_int(n, k, replace=False, accelerated=use_numba, seed=123)
-    samples_2 = sample_int(n, k, replace=False, accelerated=use_numba, seed=123)
-    samples_3 = sample_int(n, k, replace=False, accelerated=use_numba, seed=None)
+    samples_1 = sample_int(n, k, replace=False, p=p, use_numba=use_numba, seed=123)
+    samples_2 = sample_int(n, k, replace=False, p=p, use_numba=use_numba, seed=123)
+    samples_3 = sample_int(n, k, replace=False, p=p, use_numba=use_numba, seed=None)
+    samples_4 = sample_int(n, k, replace=False, p=p, use_numba=use_numba, seed=0)
 
     # --- assert ------------------------------------------
     np.testing.assert_array_equal(samples_1, samples_2)
@@ -189,6 +201,7 @@ def test_sample_int_uniform_without_replacement_seed(n: int, k: int, use_numba: 
         # in this case, it's very unlikely that samples_3 matches samples_1 or samples_2
         assert not list(samples_1) == list(samples_3)
         assert not list(samples_2) == list(samples_3)
+        assert not list(samples_3) == list(samples_4)
 
 
 # =================================================================================================
@@ -201,7 +214,7 @@ def test_sample_int_non_uniform_with_replacement_invariants_single(n: int, use_n
     p = get_probabilities(n)
 
     # --- act ---------------------------------------------
-    sample = sample_int(n, replace=True, p=p, accelerated=use_numba)
+    sample = sample_int(n, replace=True, p=p, use_numba=use_numba)
 
     # --- assert ------------------------------------------
     assert isinstance(sample, numbers.Integral)
@@ -227,7 +240,7 @@ def test_sample_int_non_uniform_with_replacement_invariants_multiple(n: int, k: 
     p = get_probabilities(n)
 
     # --- act ---------------------------------------------
-    samples = sample_int(n, k, replace=True, p=p, accelerated=use_numba)
+    samples = sample_int(n, k, replace=True, p=p, use_numba=use_numba)
 
     # --- assert ------------------------------------------
     assert isinstance(samples, np.ndarray)
@@ -242,7 +255,7 @@ def test_sample_int_non_uniform_with_replacement_duplicates(use_numba: bool):
     p = get_probabilities(1000)
 
     # --- act ---------------------------------------------
-    samples = sample_int(n=1000, k=900, replace=True, p=p, accelerated=use_numba)  # very unlikely all are unique
+    samples = sample_int(n=1000, k=900, replace=True, p=p, use_numba=use_numba)  # very unlikely all are unique
 
     # --- assert ------------------------------------------
     assert len(set(samples)) < 900  # there are duplicates, which is expected with replacement
@@ -267,9 +280,10 @@ def test_sample_int_non_uniform_with_replacement_seed(n: int, k: int, use_numba:
     p = get_probabilities(n)
 
     # --- act ---------------------------------------------
-    samples_1 = sample_int(n, k, replace=True, p=p, accelerated=use_numba, seed=123)
-    samples_2 = sample_int(n, k, replace=True, p=p, accelerated=use_numba, seed=123)
-    samples_3 = sample_int(n, k, replace=True, p=p, accelerated=use_numba, seed=None)
+    samples_1 = sample_int(n, k, replace=True, p=p, use_numba=use_numba, seed=123)
+    samples_2 = sample_int(n, k, replace=True, p=p, use_numba=use_numba, seed=123)
+    samples_3 = sample_int(n, k, replace=True, p=p, use_numba=use_numba, seed=None)
+    samples_4 = sample_int(n, k, replace=True, p=p, use_numba=use_numba, seed=0)
 
     # --- assert ------------------------------------------
     np.testing.assert_array_equal(samples_1, samples_2)
@@ -277,6 +291,7 @@ def test_sample_int_non_uniform_with_replacement_seed(n: int, k: int, use_numba:
         # in this case, it's very unlikely that samples_3 matches samples_1 or samples_2
         assert not list(samples_1) == list(samples_3)
         assert not list(samples_2) == list(samples_3)
+        assert not list(samples_3) == list(samples_4)
 
 
 @pytest.mark.parametrize("use_numba", [True, False])
@@ -292,7 +307,7 @@ def test_sample_int_non_uniform_with_replacement_probs(use_numba: bool, factor: 
     expected_mean = sum(i * p[i] for i in range(n))
 
     # --- act ---------------------------------------------
-    samples = sample_int(n=n, k=k, replace=True, p=p, accelerated=use_numba)
+    samples = sample_int(n=n, k=k, replace=True, p=p, use_numba=use_numba)
 
     # --- assert ------------------------------------------
     assert 0.95 * expected_mean < np.mean(samples) < 1.05 * expected_mean
@@ -309,7 +324,7 @@ def test_sample_int_non_uniform_without_replacement_invariants_single(n: int, us
     p = get_probabilities(n)
 
     # --- act ---------------------------------------------
-    sample = sample_int(n, replace=False, p=p, accelerated=use_numba)
+    sample = sample_int(n, replace=False, p=p, use_numba=use_numba)
 
     # --- assert ------------------------------------------
     assert isinstance(sample, numbers.Integral)
@@ -334,7 +349,7 @@ def test_sample_int_non_uniform_without_replacement_invariants_multiple(n: int, 
     p = get_probabilities(n)
 
     # --- act ---------------------------------------------
-    samples = sample_int(n, k, replace=False, p=p, accelerated=use_numba)
+    samples = sample_int(n, k, replace=False, p=p, use_numba=use_numba)
 
     # --- assert ------------------------------------------
     assert isinstance(samples, np.ndarray)
@@ -362,9 +377,9 @@ def test_sample_int_non_uniform_without_replacement_seed(n: int, k: int, use_num
     p = get_probabilities(n)
 
     # --- act ---------------------------------------------
-    samples_1 = sample_int(n, k, replace=False, p=p, accelerated=use_numba, seed=123)
-    samples_2 = sample_int(n, k, replace=False, p=p, accelerated=use_numba, seed=123)
-    samples_3 = sample_int(n, k, replace=False, p=p, accelerated=use_numba, seed=None)
+    samples_1 = sample_int(n, k, replace=False, p=p, use_numba=use_numba, seed=123)
+    samples_2 = sample_int(n, k, replace=False, p=p, use_numba=use_numba, seed=123)
+    samples_3 = sample_int(n, k, replace=False, p=p, use_numba=use_numba, seed=None)
 
     # --- assert ------------------------------------------
     np.testing.assert_array_equal(samples_1, samples_2)
@@ -387,7 +402,7 @@ def test_sample_int_non_uniform_without_replacement_probs(use_numba: bool, facto
     expected_mean = sum(i * p[i] for i in range(n))  # approx. correct; this assumes replacement
 
     # --- act ---------------------------------------------
-    samples = sample_int(n=n, k=k, replace=False, p=p, accelerated=use_numba)
+    samples = sample_int(n=n, k=k, replace=False, p=p, use_numba=use_numba)
 
     # --- assert ------------------------------------------
     assert 0.95 * expected_mean < np.mean(samples) < 1.05 * expected_mean