+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 > *(under development)*
 <!------------------------------------------------------------------------------------------------->
 
+- **new**:
+  - make `numba` and non-`numba` versions of `sampling.discrete.sample_int` public and self-sustained (full-featured)
+
 - **improved**:
   - `sampling.discrete.sample_int`: improve benchmarking + add notes on optimization efforts
   - improve collection of test coverage data for `numba`-decorated functions